feat/fix: use float32 for muon to make it compatible with v100

+++ optimizer.py
@@ -12,7 +12,7 @@ def zeropower_via_newtonschulz5(G: Tensor, steps: int) -> Tensor:
 	"""
 	assert G.ndim >= 2
 	a, b, c = (3.4445, -4.7750, 2.0315)
-	X = G.to(torch.bfloat16)
+	X = G.to(torch.float32)
 
 	if G.size(-2) > G.size(-1):
 		X = X.mT
